move search.jsx back to partials
