+++ app/src/scripts/dashboard/dashboard.datasets.jsx
@@ -18,7 +18,7 @@ import Filters from './dashboard.filters.jsx'
 import Sort from './dashboard.sort.jsx'
 import Summary from '../dataset/dataset.summary.jsx'
 import Input from '../common/forms/input.jsx'
-import Search from './dashboard.search.jsx'
+import Search from '../common/partials/search.jsx'
 
 import { refluxConnect } from '../utils/reflux'
 import { pageTitle } from '../resources/strings'
